Update README.md
Just realized the mention of GUI did not make a need for it to be in it's own header and is related to the manual command above. Unsure of how to edit my merge, but the "worry" and few additional words still stand, and a break between the mention of GUI and the font import.